Code/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
essential-generators==1.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
urllib3==1.26.8
Werkzeug==2.0.2
